refactor dialog content into new folder
makes room for upcoming resize image dialog
