+++ src/components/dialogs/app-info/AppInfo.tsx
