+++ src/components/dialogs/app-info/AppInfoProject.tsx
