+++ src/components/dialogs/app-info/sections/Contributors.tsx
