+++ src/components/dialogs/app-info/sections/GettingStarted.tsx
