+++ src/components/dialogs/app-info/sections/KeyboardShortcuts.tsx
