+++ src/components/dialogs/app-info/sections/OtherProjects.tsx
