+++ src/components/dialogs/app-info/sections/SceneNavigationHints.tsx
