+++ src/components/dialogs/index.tsx
@@ -1,0 +1,2 @@
+export { default as AppDialog } from './AppDialog';
+export { default as AppInfo } from './app-info/AppInfo';
